Restructured cli/src/commands Also, made server-related commands gated under the `server` feature.
