+++ cli/Cargo.toml
@@ -40,6 +40,9 @@ tempfile.workspace = true
 regex = "1"
 serial_test = "3"
 
+[features]
+server = []
+
 [lib]
 name = "gitehr"
 path = "src/lib.rs"
