+++ cli/src/commands/server/user/create.rs
@@ -1,0 +1,5 @@
+use anyhow::Result;
+
+pub fn run() -> Result<()> {
+    crate::commands::contributor::create_user_interactive()
+}
