+++ cli/src/commands/server/user/deactivate.rs
@@ -1,0 +1,5 @@
+use anyhow::Result;
+
+pub fn run() -> Result<()> {
+    crate::commands::contributor::deactivate_contributor()
+}
