+++ cli/src/commands/server/user/disable.rs
@@ -1,0 +1,5 @@
+use anyhow::Result;
+
+pub fn run(id: &str) -> Result<()> {
+    crate::commands::contributor::disable_contributor(id)
+}
