+++ cli/src/commands/server/user/enable.rs
@@ -1,0 +1,5 @@
+use anyhow::Result;
+
+pub fn run(id: &str) -> Result<()> {
+    crate::commands::contributor::enable_contributor(id)
+}
