+++ cli/src/commands/server/user/list.rs
@@ -1,0 +1,5 @@
+use anyhow::Result;
+
+pub fn run() -> Result<()> {
+    crate::commands::contributor::list_contributors()
+}
